// src/types/stock/StockProfile.ts
export interface StockProfile {
    symbol: string;
    price: number;
    beta: number;
    lastDividend: number;
    range: string;
    change: number;
    changePercentage: number;
    companyName: string;
    currency: string;
    industry: string;
    website: string;
    description: string;
    ceo: string;
    sector: string;
    image: string;
    marketCap: number;
    fullTimeEmployees: string;
    exchange: string;
}